requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
cffi==1.15.1
cryptography==38.0.4
Django==4.1.4
django-constance==2.9.1
django-picklefield==3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.25.0
inflection==0.5.1
jsonschema==4.17.3
mutagen==1.46.0
paho-mqtt==1.6.1
pycodestyle==2.10.0
pycparser==2.21
pygame==2.1.2
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
